nlp/CH4/practice/requirements.txt
qdrant-client==1.7.3
sentence-transformers==2.5.1
fastapi==0.109.2
uvicorn==0.27.1
datasets==2.17.0
openai==1.12.0
python-dotenv==1.0.1
scikit-learn==1.4.0
numpy==1.26.4